django_backend/requirements.txt

Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
